feat: Changed controller to protocol
